requirements.txt
numpy
pandas
scikit-learn
nltk
flask
plotly
dash
pulp
python-dotenv
yfinance
requests
wbgapi
beautifulsoup4
html5lib
spacy
scipy
seaborn
matplotlib
dash-bootstrap-components
